Share display text truncation helpers
