+++ agent-cli-wrapper/claude/render/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
     ],
     importpath = "github.com/bazelment/yoloswe/agent-cli-wrapper/claude/render",
     visibility = ["//visibility:public"],
+    deps = ["//agent-cli-wrapper/displaytext"],
 )
 
 go_test(
