+++ agent-cli-wrapper/displaytext/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "displaytext",
+    srcs = ["displaytext.go"],
+    importpath = "github.com/bazelment/yoloswe/agent-cli-wrapper/displaytext",
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "displaytext_test",
+    srcs = ["displaytext_test.go"],
+    deps = [":displaytext"],
+)
