+++ bramble/sessionmodel/format.go
@@ -2,7 +2,8 @@ package sessionmodel
 
 import (
 	"fmt"
-	"strings"
+
+	"github.com/bazelment/yoloswe/agent-cli-wrapper/displaytext"
 )
 
 // FormatToolContent creates a display-friendly content string for a tool call.
@@ -15,15 +16,15 @@ func FormatToolContent(name string, input map[string]interface{}) string {
 	switch name {
 	case "Read":
 		if path, ok := input["file_path"].(string); ok {
-			return fmt.Sprintf("%s %s", name, truncatePathForDisplay(path))
+			return fmt.Sprintf("%s %s", name, displaytext.TruncatePath(path, 60))
 		}
 	case "Write", "Edit":
 		if path, ok := input["file_path"].(string); ok {
-			return fmt.Sprintf("%s → %s", name, truncatePathForDisplay(path))
+			return fmt.Sprintf("%s → %s", name, displaytext.TruncatePath(path, 60))
 		}
 	case "Bash":
 		if cmd, ok := input["command"].(string); ok {
-			return fmt.Sprintf("%s: %s", name, TruncateForDisplay(cmd, 50))
+			return fmt.Sprintf("%s: %s", name, displaytext.Truncate(cmd, 50))
 		}
 	case "Glob":
 		if pattern, ok := input["pattern"].(string); ok {
@@ -31,43 +32,12 @@ func FormatToolContent(name string, input map[string]interface{}) string {
 		}
 	case "Grep":
 		if pattern, ok := input["pattern"].(string); ok {
-			return fmt.Sprintf("%s %s", name, TruncateForDisplay(pattern, 40))
+			return fmt.Sprintf("%s %s", name, displaytext.Truncate(pattern, 40))
 		}
 	case "Task":
 		if desc, ok := input["description"].(string); ok {
-			return fmt.Sprintf("%s: %s", name, TruncateForDisplay(desc, 40))
+			return fmt.Sprintf("%s: %s", name, displaytext.Truncate(desc, 40))
 		}
 	}
 	return name
-}
-
-// truncatePathForDisplay keeps the filename visible when truncating paths.
-func truncatePathForDisplay(path string) string {
-	runes := []rune(path)
-	if len(runes) <= 60 {
-		return path
-	}
-	if lastSlash := strings.LastIndexByte(path, '/'); lastSlash > 0 {
-		suffix := path[lastSlash:]
-		if len([]rune(suffix)) <= 50 {
-			return "..." + suffix
-		}
-	}
-	return string(runes[:57]) + "..."
-}
-
-// TruncateForDisplay truncates s to at most max Unicode code points, appending
-// "..." if truncation occurred (the suffix counts toward max). Rune-based
-// indexing avoids splitting multi-byte UTF-8 sequences that byte-based slicing
-// would corrupt. If max is less than 3, the string is truncated to max runes
-// with no suffix.
-func TruncateForDisplay(s string, max int) string {
-	runes := []rune(s)
-	if len(runes) <= max {
-		return s
-	}
-	if max <= 3 {
-		return string(runes[:max])
-	}
-	return string(runes[:max-3]) + "..."
 }
