+++ bramble/sessionmodel/format_test.go
@@ -1,0 +1,15 @@
+package sessionmodel
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestFormatToolContentTruncatesLongPathWithParent(t *testing.T) {
+	content := FormatToolContent("Read", map[string]interface{}{
+		"file_path": "/really/long/prefix/that/exceeds/sixty/chars/with/many/segments/deeply/nested/file.go",
+	})
+
+	assert.Equal(t, "Read .../nested/file.go", content)
+}
